+++ python/assemble.sh
@@ -24,11 +24,13 @@ cargo run -p nostr-sdk-ffi --features uniffi-cli --bin uniffi-bindgen generate -
 # Build linux glibc wheels
 docker run --rm -v "${TARGET_DIR}/i686-unknown-linux-gnu/release/libnostr_sdk_ffi.so:/build/binaries/libnostr_sdk_ffi.so" -v "${SRC_DIR}:/build/binding" -v "$(pwd)/dist:/build/dist" -e PLAT_NAME="manylinux_2_17_i686" wheel-builder
 docker run --rm -v "${TARGET_DIR}/x86_64-unknown-linux-gnu/release/libnostr_sdk_ffi.so:/build/binaries/libnostr_sdk_ffi.so" -v "${SRC_DIR}:/build/binding" -v "$(pwd)/dist:/build/dist" -e PLAT_NAME="manylinux_2_17_x86_64" wheel-builder
+docker run --rm -v "${TARGET_DIR}/armv7-unknown-linux-gnueabihf/release/libnostr_sdk_ffi.so:/build/binaries/libnostr_sdk_ffi.so" -v "${SRC_DIR}:/build/binding" -v "$(pwd)/dist:/build/dist" -e PLAT_NAME="manylinux_2_17_armv7l" wheel-builder
 docker run --rm -v "${TARGET_DIR}/aarch64-unknown-linux-gnu/release/libnostr_sdk_ffi.so:/build/binaries/libnostr_sdk_ffi.so" -v "${SRC_DIR}:/build/binding" -v "$(pwd)/dist:/build/dist" -e PLAT_NAME="manylinux_2_17_aarch64" wheel-builder
 
 # Build linux musl wheels
 docker run --rm -v "${TARGET_DIR}/i686-unknown-linux-musl/release/libnostr_sdk_ffi.so:/build/binaries/libnostr_sdk_ffi.so" -v "${SRC_DIR}:/build/binding" -v "$(pwd)/dist:/build/dist" -e PLAT_NAME="musllinux_1_2_i686" wheel-builder
 docker run --rm -v "${TARGET_DIR}/x86_64-unknown-linux-musl/release/libnostr_sdk_ffi.so:/build/binaries/libnostr_sdk_ffi.so" -v "${SRC_DIR}:/build/binding" -v "$(pwd)/dist:/build/dist" -e PLAT_NAME="musllinux_1_2_x86_64" wheel-builder
+docker run --rm -v "${TARGET_DIR}/armv7-unknown-linux-musleabihf/release/libnostr_sdk_ffi.so:/build/binaries/libnostr_sdk_ffi.so" -v "${SRC_DIR}:/build/binding" -v "$(pwd)/dist:/build/dist" -e PLAT_NAME="musllinux_1_2_armv7l" wheel-builder
 docker run --rm -v "${TARGET_DIR}/aarch64-unknown-linux-musl/release/libnostr_sdk_ffi.so:/build/binaries/libnostr_sdk_ffi.so" -v "${SRC_DIR}:/build/binding" -v "$(pwd)/dist:/build/dist" -e PLAT_NAME="musllinux_1_2_aarch64" wheel-builder
 
 # Build macos wheels