+++ scripts/linux.sh
@@ -19,9 +19,11 @@ pip install cargo-zigbuild==0.19.8
 # Build (GLIBC 2.17)
 cargo zigbuild -p nostr-sdk-ffi --target i686-unknown-linux-gnu.2.17 --release
 cargo zigbuild -p nostr-sdk-ffi --target x86_64-unknown-linux-gnu.2.17 --release
+cargo zigbuild -p nostr-sdk-ffi --target armv7-unknown-linux-gnueabihf.2.17 --release
 cargo zigbuild -p nostr-sdk-ffi --target aarch64-unknown-linux-gnu.2.17 --release
 
 # Build (MUSL)
 cargo zigbuild -p nostr-sdk-ffi --target i686-unknown-linux-musl --release
 cargo zigbuild -p nostr-sdk-ffi --target x86_64-unknown-linux-musl --release
+cargo zigbuild -p nostr-sdk-ffi --target armv7-unknown-linux-musleabihf --release
 cargo zigbuild -p nostr-sdk-ffi --target aarch64-unknown-linux-musl --release